adc: pulsar_adc: cast rx buffer to rx_addr integer
spi_engine_offload_message.rx_addr is a uint32_t, but the offload read
assigned the uint32_t *buf pointer directly, which -Wint-conversion
flags. Cast to uint32_t, matching the ad469x offload idiom.

+++ drivers/adc/pulsar_adc/pulsar_adc.c
@@ -199,7 +199,7 @@ static int32_t pulsar_adc_read_data_offload(struct pulsar_adc_dev *dev,
 
 	msg.commands = spi_eng_msg_cmds;
 	msg.no_commands = NO_OS_ARRAY_SIZE(spi_eng_msg_cmds);
-	msg.rx_addr = buf;
+	msg.rx_addr = (uint32_t)buf;
 	msg.commands_data = commands_data;
 
 	ret = spi_engine_offload_transfer(dev->spi_desc, msg, samples);
